refactor(setting-tabs): move all tabs from pages to SafeSettings directory; move AddCreateProposalPermissionModal under ui/modals
