+++ src/components/SafeSettings/TabContents/governance/SafeGovernanceSettingTab.tsx
@@ -1,25 +1,19 @@
-import { Box, Show, Text } from '@chakra-ui/react';
+import { Box, Text } from '@chakra-ui/react';
 import { useTranslation } from 'react-i18next';
-import { zeroAddress } from 'viem';
-import { ERC20TokenContainer } from '../../../../components/SafeSettings/ERC20TokenContainer';
-import { ERC721TokensContainer } from '../../../../components/SafeSettings/ERC721TokensContainer';
-import { SettingsContentBox } from '../../../../components/SafeSettings/SettingsContentBox';
-import { SignersContainer } from '../../../../components/SafeSettings/Signers/SignersContainer';
-import NestedPageHeader from '../../../../components/ui/page/Header/NestedPageHeader';
-import { DAO_ROUTES } from '../../../../constants/routes';
 import { useCurrentDAOKey } from '../../../../hooks/DAO/useCurrentDAOKey';
 import { useDAOStore } from '../../../../providers/App/AppProvider';
-import { useNetworkConfigStore } from '../../../../providers/NetworkConfig/useNetworkConfigStore';
 import { GovernanceType } from '../../../../types';
+import { ERC20TokenContainer } from '../../ERC20TokenContainer';
+import { ERC721TokensContainer } from '../../ERC721TokensContainer';
+import { SettingsContentBox } from '../../SettingsContentBox';
+import { SignersContainer } from '../../Signers/SignersContainer';
 import { GovernanceParams } from './GovernanceParams';
 
-export function SafeGovernanceSettingsPage() {
+export function SafeGovernanceSettingTab() {
   const { t } = useTranslation('settings');
-  const { addressPrefix } = useNetworkConfigStore();
   const { daoKey } = useCurrentDAOKey();
   const {
     governance: { type },
-    node: { safe },
   } = useDAOStore({ daoKey });
 
   const isERC20Governance = type === GovernanceType.AZORIUS_ERC20;
@@ -28,15 +22,6 @@ export function SafeGovernanceSettingsPage() {
 
   return (
     <>
-      <Show below="md">
-        <NestedPageHeader
-          title={t('daoSettingsGovernance')}
-          backButton={{
-            text: t('settings'),
-            href: DAO_ROUTES.settings.relative(addressPrefix, safe?.address || zeroAddress),
-          }}
-        />
-      </Show>
       <SettingsContentBox
         display="flex"
         flexDirection="column"
