+++ src/constants/routes.ts
@@ -97,41 +97,6 @@ export const DAO_ROUTES = {
       `/proposals/new/sablier${getDaoQueryParam(addressPrefix, daoAddress)}`,
     path: 'proposals/new/sablier',
   },
-  settings: {
-    relative: (addressPrefix: string, safeAddress: string) =>
-      `/settings${getDaoQueryParam(addressPrefix, safeAddress)}`,
-    path: 'settings',
-  },
-  settingsGeneral: {
-    relative: (addressPrefix: string, daoAddress: string) =>
-      `/settings/general${getDaoQueryParam(addressPrefix, daoAddress)}`,
-    path: 'general',
-  },
-  settingsGovernance: {
-    relative: (addressPrefix: string, daoAddress: string) =>
-      `/settings/governance${getDaoQueryParam(addressPrefix, daoAddress)}`,
-    path: 'governance',
-  },
-  settingsModulesAndGuard: {
-    relative: (addressPrefix: string, daoAddress: string) =>
-      `/settings/modules-and-guard${getDaoQueryParam(addressPrefix, daoAddress)}`,
-    path: 'modules-and-guard',
-  },
-  settingsPermissions: {
-    relative: (addressPrefix: string, daoAddress: string) =>
-      `/settings/permissions${getDaoQueryParam(addressPrefix, daoAddress)}`,
-    path: 'permissions',
-  },
-  settingsPermissionsCreateProposal: {
-    relative: (addressPrefix: string, daoAddress: string, votingStrategyAddress?: string) =>
-      `/settings/permissions/create-proposal${getDaoQueryParam(addressPrefix, daoAddress)}${votingStrategyAddress ? `&votingStrategy=${votingStrategyAddress}` : ''}`,
-    path: 'create-proposal',
-  },
-  settingsToken: {
-    relative: (addressPrefix: string, daoAddress: string) =>
-      `/settings/token${getDaoQueryParam(addressPrefix, daoAddress)}`,
-    path: 'token',
-  },
   proposalTemplates: {
     relative: (addressPrefix: string, safeAddress: string) =>
       `/proposal-templates${getDaoQueryParam(addressPrefix, safeAddress)}`,
